build: update package references
