+++ package.json
@@ -14,7 +14,7 @@
     "preview": "vite preview"
   },
   "devDependencies": {
-    "@biomejs/biome": "^1.9.4",
+    "@biomejs/biome": "^2.3.7",
     "@testing-library/jest-dom": "^6.6.3",
     "@testing-library/react": "^16.3.0",
     "@testing-library/user-event": "^14.6.1",
@@ -22,15 +22,15 @@
     "@types/node": "^24.0.3",
     "@types/react": "^19.1.6",
     "@types/react-dom": "^19.1.6",
-    "@vitejs/plugin-react-swc": "^3.10.1",
-    "@vitest/coverage-v8": "^3.2.2",
-    "@vitest/ui": "^3.2.2",
+    "@vitejs/plugin-react-swc": "^4.2.2",
+    "@vitest/coverage-v8": "^4.0.13",
+    "@vitest/ui": "^4.0.13",
     "happy-dom": "^20.0.10",
-    "lefthook": "^1.11.13",
+    "lefthook": "^2.0.4",
     "typescript": "^5.8.3",
-    "vite": "^6.4.1",
+    "vite": "^7.2.4",
     "vite-tsconfig-paths": "^5.1.4",
-    "vitest": "^3.2.2"
+    "vitest": "^4.0.13"
   },
   "dependencies": {
     "@emotion/react": "^11.14.0",
@@ -44,7 +44,7 @@
     "react": "^19.1.0",
     "react-ace": "^14.0.1",
     "react-dom": "^19.1.0",
-    "react-i18next": "^15.5.2",
+    "react-i18next": "^16.3.5",
     "terser": "^5.42.0",
     "vite-plugin-svgr": "^4.5.0"
   }
