--- src/__mocks__/emotion.ts
@@ -1,1 +1,0 @@
-export const css = () => ''
