--- src/__mocks__/svg.tsx
@@ -1,2 +1,0 @@
-const Dummy = () => <span />
-export default Dummy
