+++ src/views/atoms/__mocks__/P5JsSandBox.tsx
@@ -1,0 +1,3 @@
+import type { JSX } from 'react'
+
+export const P5JsSandBox: React.FC = (): JSX.Element => <div title="canvas" />
