+++ vitest.config.ts
@@ -1,47 +1,29 @@
 // @ts-nocheck
-import * as path from 'node:path'
 import react from '@vitejs/plugin-react-swc'
+import svgr from 'vite-plugin-svgr'
 import tsconfigPaths from 'vite-tsconfig-paths'
 import { defineConfig } from 'vitest/config'
-
-import svgr from 'vite-plugin-svgr'
 
 export default defineConfig({
   plugins: [react(), tsconfigPaths(), svgr()],
   test: {
-    globals: true,
-    root: 'src',
     environment: 'happy-dom',
-    setupFiles: './src/setupTest.ts',
-    aliases: [
-      {
-        find: /.*\.svg/,
-        replacement: path.resolve(__dirname, './src/__mock__/svg.tsx'),
-      },
-      {
-        find: /@emotion\/react/,
-        replacement: path.resolve(__dirname, './src/__mocks__/emotion.ts'),
-      },
-      {
-        find: /.*\/GlobalStyles/,
-        replacement: path.resolve(__dirname, './src/__mocks__/GlobalStyles.tsx'),
-      },
-      {
-        find: /.*\/CodeEditor/,
-        replacement: path.resolve(__dirname, './src/__mocks__/CodeEditor.tsx'),
-      },
-    ],
+    globals: true,
+    root: '',
+    setupFiles: 'src/setupTest.ts',
     coverage: {
       exclude: [
         '__mocks__',
         '**/types.ts',
-        './vite-env.d.ts',
-        './setupTest.d.ts',
-        './main.tsx',
-        'views/GlobalStyles.tsx',
+        'src/vite-env.d.ts',
+        'src/setupTest.d.ts',
+        'src/main.tsx',
+        'src/views/GlobalStyles.tsx',
+        'src/views/atoms/P5JsSandBox.tsx',
+        'src/views/molecules/CodeEditor.tsx',
       ],
       reporter: ['text', 'json', 'html'],
-      reportsDirectory: '../build/coverage',
+      reportsDirectory: 'build/coverage',
     },
   },
 })
